Applied suggested fixes from PR
