+++ README.md
@@ -1,20 +1,77 @@
-### Your First Express and React App
-Create full stack apps with React and Express. Run your client, server and do a restart of your dump db to have a full-stack project working in minutes
+# ISBN Thinking
 
-# Step by Step instructions - To use this project as your start point
+ISBN Thinking is a library and book list manager. Created as a way to keep myself from getting 4 copies of the same book and keep track of titles that look interesting.
+
+## Description
+
+## Pre-requirements
+
+Docker
+
+## Installation
+
+### To run locally:
+
+## Set Up the Development Environment
+
+### Install NPM Packages
+
+```
+npm install
+```
+
+### Set Up `postgres` User Password and Database Name
+
+We need to set up couple pieces of information in order to start a new
+PostgreSQL server instance, as well as to connect to it later from the Express
+server.
+
+1. Copy the example environment file
+
+   ```sh
+   cp .env.example .env
+   ```
+
+2. You can choose to edit `.env` or just use as-is.
+
+[See the PostgreSQL Docker image documentation for more
+information][dh-postgres].
+
+### Initialize the Database
+
+Let's set up the database server, create the application database, and seed it
+with some data. You only need to do this the first time you set up your
+development environment.
+
+```sh
+npm run db:init
+```
+
+ℹ️ If you ever need to start over with the database, you can run this command
+again which will delete your existing data and start from scratch.
+
+## Start the Development Environment
+
+```sh
+npm start
+```
+
+Visit <http://localhost:3000>.
+
+## Usage
+
 ### For create the whole project
-1. Go to your source directory in your terminal and run the command `git clone https://github.com/Yosolita1978/React-Express-PairProgramming.git NAMENEWDIRECTORY`
-[!You will something like this in your terminal.](https://github.com/Yosolita1978/screenshoots/blob/main/template/Screen%20Shot%202022-03-20%20at%207.50.46%20PM.png?raw=true)
 
-2. To restore the DB dump file that the project already contain, just run the command `psql -U postgres -f db.sql`. Make sure that you have your Postgres password on hand. The psql console will ask you for your password. 
+1. Go to your source directory in your terminal and run the command `git clone https://github.com/Yosolita1978/React-Express-PairProgramming.git NAMENEWDIRECTORY`
+
+2. To restore the DB dump file that the project already contain, just run the command `psql -U postgres -f db.sql`. Make sure that you have your Postgres password on hand. The psql console will ask you for your password.
 3. To clean your folder from the owner git, run the command `rm -rf .git`
 4. Run the command `git init` to start your git repository
 5. Go to the server folder in the project (`cd server`) and run the command `npm install`
 6. Inside your server folder, create an .env file with `touch .env`
-7. Inside your server folder, open the file `.env.example` and copy the file there. 
+7. Inside your server folder, open the file `.env.example` and copy the file there.
 8. Inside your .env file, paste the string from .env.example and change the variables with the values from the project. For this template, don't change the name of your db.
 9. Inside your server file: run the command `psql -U postgres -f db.sql` to restore the DB from the file db.sql
 10. Go to the cliente folder (`cd .. and cd client`) and run the command `npm start`
 11. Both server should run now with `npm start`
 12. Go to localhost:3000 and you should see something like this
-[!You will something like this in your terminal.](https://github.com/Yosolita1978/screenshoots/blob/main/template/Screen%20Shot%202022-03-20%20at%208.58.13%20PM.png?raw=true)
