+++ client/src/App.css
@@ -1,16 +1,11 @@
 .App {
   text-align: center;
+  font-family: "Marcellus", serif;
 }
 
 .App-logo {
   height: 40vmin;
   pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
 }
 
 .App-header {
@@ -28,15 +23,22 @@
   color: #61dafb;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+ul {
+  list-style: none;
+}
+.component-page {
+  display: "flex";
+  justify-content: "Right";
+  height: "100vh";
+  flex-direction: "column";
+  align-items: "center";
 }
 
-ul{
-  list-style: none;
+.individual-book {
+  display: "flex";
+  align-items: "center";
+}
+
+fieldset.radio-form {
+  display: "inline";
 }
