+++ client/src/App.js
@@ -1,12 +1,11 @@
 import "./App.css";
-import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
-import BookList from "./components/book_list";
+import "./pages/page.css";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import Navbar from "./components/Navbar";
 import Home from "./pages";
 import Books from "./pages/books";
 import AddNewBook from "./pages/add_new_book";
 import SearchBook from "./pages/search_book";
-import AddApiBook from "./pages/add_api_book";
 
 function App() {
   return (
@@ -14,13 +13,10 @@ function App() {
       <Navbar />
 
       <div>
-        {/* A <Switch> looks through its children <Route>s and
-            renders the first one that matches the current URL. */}
         <Switch>
           <Route path="/" exact component={Home} />
           <Route path="/books" component={Books} />
           <Route path="/addnewbook" component={AddNewBook} />
-          <Route path="/apibook" component={AddApiBook} />
           <Route path="/search" component={SearchBook} />
         </Switch>
       </div>
