+++ client/src/components/Navbar/index.js
@@ -1,70 +1,18 @@
 import React from "react";
-import {
-  Nav,
-  NavLink,
-  Bars,
-  NavMenu,
-  NavBtn,
-  NavBtnLink,
-} from "./NavbarElements";
+import { Nav, NavLink, NavMenu } from "./NavbarElements";
 
 const Navbar = () => {
   return (
     <>
       <Nav>
-        <Bars />
-
         <NavMenu>
-          <NavLink
-            exact
-            to="/"
-            activeStyle={{
-              fontWeight: "bold",
-              color: "darkgreen",
-            }}
-          >
+          <NavLink exact to="/">
             Home
           </NavLink>
-          <NavLink
-            to="/books"
-            activeStyle={{
-              fontWeight: "bold",
-              color: "darkgreen",
-            }}
-          >
-            Books
-          </NavLink>
-          <NavLink
-            to="/addnewbook"
-            activeStyle={{
-              fontWeight: "bold",
-              color: "darkgreen",
-            }}
-          >
-            Add A Book
-          </NavLink>
-          <NavLink
-            to="/apibook"
-            activeStyle={{
-              fontWeight: "bold",
-              color: "darkgreen",
-            }}
-          >
-            Add API Book
-          </NavLink>
-          <NavLink
-            to="/search"
-            activeStyle={{
-              fontWeight: "bold",
-              color: "darkgreen",
-            }}
-          >
-            Search
-          </NavLink>
-          {/* Second Nav */}
-          {/* <NavBtnLink to='/sign-in'>Sign In</NavBtnLink> */}
+          <NavLink to="/books">Books</NavLink>
+          <NavLink to="/addnewbook">Add A Book</NavLink>
+          <NavLink to="/search">Search</NavLink>
         </NavMenu>
-        <NavBtn></NavBtn>
       </Nav>
     </>
   );
