+++ client/src/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
+import "./pages/page.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
