+++ client/src/pages/page.css
@@ -1,0 +1,16 @@
+.component-page {
+  display: "flex";
+  justify-content: "Right";
+  height: "100vh";
+  flex-direction: "column";
+  align-items: "center";
+}
+
+.individual-book {
+  display: "flex";
+  align-items: "center";
+}
+
+fieldset.radio-form {
+  display: "inline";
+}
